lab1/requirements.txt
apache-airflow==2.7.1
apache-airflow-providers-postgres
pendulum
pandas
numpy
scikit-learn